backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.33.3
python-dotenv==1.0.0
PyJWT==2.10.1
passlib==1.7.4
supabase==2.24.0
gunicorn==21.2.0  # For production